requirements.txt
beautifulsoup4==4.12.2
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.1
selenium==4.10.0
selenium-requests==2.0.3
webdriver-manager==4.0.1